Improve base-10 itoa implementation.
Modified the base-10 itoa replacement to use an std::basic_ostringstream
and return a tstring. This enabled removal of a bunch of character
arrays used as buffers for the conversions. Also, renamed function from
itot10 to itoa10. I have no idea why I didn't call it itoa10 in the
first place.